fix: remove placeholder text from addon modules
