+++ addons/blender_tools/__init__.py
@@ -1,0 +1,19 @@
+bl_info = {
+    "name": "Blender Tools",
+    "blender": (4, 3, 2),
+    "category": "Object",
+}
+
+__version__ = "0.1.0"
+
+
+def register() -> None:
+    from . import operators
+
+    operators.register()
+
+
+def unregister() -> None:
+    from . import operators
+
+    operators.unregister()
